Optionally enable file injection
There is a patch up for nova right now that disables file injection by default.
This is a corresponding devstack change that only sets file injection options if
it is enabled in the devstack config.  This is good to keep around so that we
can easily turn it on for testing.

The nova change is id Icff1304fc816acc843f8962727aef8bbbc7bbaa3.

Change-Id: I5015f2c351b1d680c205d7f9a5204febca490b91