Set uri_v3 in tempest config
This properly sets the v3 uri for keystone in the tempest config.
Previously tempest would just guess the v3 uri by replacing v2 with v3.
However, moving forward this will no longer be the case so devstack
should properly set this uri to enable tempest to use the keystone v3
api in addition to the v2.

Change-Id: Ib02b2e9f24d8ca1f381186c48747ca0fbc45f3f1