Stop catting tempest.config during tempest setup
This commit removes the 'cat tempest.config' from lib/tempest. There
is no reason to cat it as part of running devstack because the file
is and can be interacted with after devstack finishes running. To
prevent a loss of information in the gate this change should be
coupled with devstack-gate change: Ifb36918cd0d686cb3865f5322cd62c209acaaf30
which copies the tempest.config file with the other test artifacts.

Change-Id: Ia01cd53660b3490ea9faa9e9c746bafd9df12a9b